Add -Xsource:3 flag and fix errors
