+++ core/src/main/scala/zongo/Mongo.scala
@@ -21,15 +21,19 @@ object Mongo {
 
     /** Gets the database with the given name.
       *
-      * @param name the name of the database
-      * @return the database wrapped in a ZIO.
+      * @param name
+      *   the name of the database
+      * @return
+      *   the database wrapped in a ZIO.
       */
     def getDatabase(name: String): Task[MongoDatabase]
 
     /** Drops the database.
       *
-      * @param db the database to drop.
-      * @return unit.
+      * @param db
+      *   the database to drop.
+      * @return
+      *   unit.
       */
     def dropDatabase(db: MongoDatabase): Task[Unit]
 
@@ -46,19 +50,25 @@ object Mongo {
 
     /** Executes command in the context of the current database.
       *
-      * @param command        the command to be run
-      * @param db             the database to use.
-      * @return a ZIO Stream containing the command result.
+      * @param command
+      *   the command to be run
+      * @param db
+      *   the database to use.
+      * @return
+      *   a ZIO Stream containing the command result.
       */
     def runCommand(command: Bson)(db: MongoDatabase): Task[Document]
 
     /** Executes command in the context of the current database.
       *
-      * @param command        the command to be run
-      * @param readPreference the ReadPreference to be used when executing
-      *                       the command
-      * @param db             the database to use.
-      * @return a ZIO Stream containing the command result.
+      * @param command
+      *   the command to be run
+      * @param readPreference
+      *   the ReadPreference to be used when executing the command
+      * @param db
+      *   the database to use.
+      * @return
+      *   a ZIO Stream containing the command result.
       */
     def runCommand(
         command: Bson,
@@ -67,25 +77,32 @@ object Mongo {
 
     /** Same as ping.
       *
-      * @param db   the database to use.
-      * @return nothing useful.
+      * @param db
+      *   the database to use.
+      * @return
+      *   nothing useful.
       */
     def healthcheck(db: MongoDatabase): Task[Unit] =
       ping(db).map(_ => ())
 
     /** Runs a query and if no error is returned all is good.
       *
-      * @param db the database to use.
-      * @return pong
+      * @param db
+      *   the database to use.
+      * @return
+      *   pong
       */
     def ping(db: MongoDatabase): Task[String] =
       runCommand(Document("ping" -> "ping"))(db).map(_ => "pong")
 
     /** Create a mongo collection.
       *
-      * @param name the name of the mongo collection to create.
-      * @param db   the database to use.
-      * @return unit.
+      * @param name
+      *   the name of the mongo collection to create.
+      * @param db
+      *   the database to use.
+      * @return
+      *   unit.
       */
     def createCollection(name: String)(
         db: MongoDatabase
@@ -94,9 +111,12 @@ object Mongo {
 
     /** Create a chunk of mongo collections.
       *
-      * @param name the names of the mongo collections to create.
-      * @param db   the database to use.
-      * @return unit.
+      * @param name
+      *   the names of the mongo collections to create.
+      * @param db
+      *   the database to use.
+      * @return
+      *   unit.
       */
     def createCollections(names: Chunk[String])(
         db: MongoDatabase
@@ -105,17 +125,23 @@ object Mongo {
 
     /** Removes ALL records from a mongo collection.
       *
-      * @param name the names of the mongo collections to create.
-      * @param db   the database to use.
-      * @return unit.
+      * @param name
+      *   the names of the mongo collections to create.
+      * @param db
+      *   the database to use.
+      * @return
+      *   unit.
       */
     def clearCollection[A](c: MongoCollection[A]): Task[DeleteResult]
 
     /** Removes ALL records from chunk of mongo collection.
       *
-      * @param name the names of the mongo collections to create.
-      * @param db   the database to use.
-      * @return unit.
+      * @param name
+      *   the names of the mongo collections to create.
+      * @param db
+      *   the database to use.
+      * @return
+      *   unit.
       */
     def clearCollections[A](
         cs: Chunk[MongoCollection[A]]
@@ -124,9 +150,12 @@ object Mongo {
 
     /** Gets the Mongo collection to use.
       *
-      * @param name the name of the mongo collection to fetch.
-      * @param db   the database to use.
-      * @return a MongoCollection instance.
+      * @param name
+      *   the name of the mongo collection to fetch.
+      * @param db
+      *   the database to use.
+      * @return
+      *   a MongoCollection instance.
       */
     def getCollection(
         name: String
@@ -134,9 +163,12 @@ object Mongo {
 
     /** Gets the Mongo collection to use for some Type.
       *
-      * @param name the name of the mongo collection to fetch.
-      * @param db   the database to use.
-      * @return a MongoCollection instance.
+      * @param name
+      *   the name of the mongo collection to fetch.
+      * @param db
+      *   the database to use.
+      * @return
+      *   a MongoCollection instance.
       */
     def getCollection[A: ClassTag](
         name: String,
@@ -145,9 +177,12 @@ object Mongo {
 
     /** Gets the Mongo collections.
       *
-      * @param name the chunk of names to fetch.
-      * @param db   the database to use.
-      * @return a MongoCollection instance.
+      * @param name
+      *   the chunk of names to fetch.
+      * @param db
+      *   the database to use.
+      * @return
+      *   a MongoCollection instance.
       */
     def getCollections(
         names: Chunk[String]
@@ -156,17 +191,23 @@ object Mongo {
 
     /** Drops a mongo collection from database.
       *
-      * @param c  the mongo collection to drop.
-      * @param db the database to use.
-      * @return unit.
+      * @param c
+      *   the mongo collection to drop.
+      * @param db
+      *   the database to use.
+      * @return
+      *   unit.
       */
     def dropCollection[A](c: MongoCollection[A]): Task[Unit]
 
     /** Drops a chunk of mongo collections from database.
       *
-      * @param cs the chunk of collections to drop.
-      * @param db the database to use.
-      * @return unit.
+      * @param cs
+      *   the chunk of collections to drop.
+      * @param db
+      *   the database to use.
+      * @return
+      *   unit.
       */
     def dropCollections[A](cs: Chunk[MongoCollection[A]]): Task[Unit] =
       ZIO.foreachPar(cs)(dropCollection(_)).map(_ => ())
@@ -175,8 +216,10 @@ object Mongo {
 
   /** Constructs a layer from a MongoClient.
     *
-    * @param uri to connect.
-    * @return a ZLayer.
+    * @param uri
+    *   to connect.
+    * @return
+    *   a ZLayer.
     */
   def live(uri: String): ULayer[Mongo] =
     MongoLive(uri).toLayer.orDie
